File and folder movement script
This script was created as a temporary solution to move folders among the teams as part of issue resolution and moving to another person's bucket

A sample scenario:
The customer support person assigns you a ticket with the respective folder containing all the documents. You make the necessary changes and then assign the ticket in the JIRA bucket, update the description and do all other things. However in most of the cases, the documentation is kept till last and finally the folder has to be moved to the respective other person's bucket (QA/Peer Review/Other team). And this is the one thing which usually gets missed. 
Sample Usage:--
This scripts first searches for a file in a location recursively. Once the file is found in a folder, then the user is asked to give the location where he would want to copy the folder. Once that is given, the complete folder is moved to the destination location with all the files and subdirectories in it.